# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Skykin platform API — developer portal (SDK keys), ad campaign portal (advertisers/operators), SDK event ingestion, intent prediction, campaign ad delivery via WebSocket, and reward notifications.",
        "title": "Skykin Platform API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/ad-portal/campaigns": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ad Portal - Campaigns"
                ],
                "summary": "List campaigns",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_campaigns_interfaces_http.CampaignListResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "One campaign row includes targeting, budget caps, and creative fields per schema. Starts inactive until activated.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ad Portal - Campaigns"
                ],
                "summary": "Create campaign with embedded creative",
                "parameters": [
                    {
                        "description": "Campaign + creative",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_campaigns_interfaces_http.CreateCampaignRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_campaigns_model.Campaign"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_platform_http.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_platform_http.APIError"
                        }
                    }
                }
            }
        },
        "/ad-portal/campaigns/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ad Portal - Campaigns"
                ],
                "summary": "Get campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_campaigns_model.Campaign"
                        }
                    }
                }
            }
        },
        "/ad-portal/campaigns/{id}/activate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sets is_active=true when validation_status is passed.",
                "tags": [
                    "Ad Portal - Campaigns"
                ],
                "summary": "Activate campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_campaigns_model.Campaign"
                        }
                    }
                }
            }
        },
        "/ad-portal/campaigns/{id}/preview": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ad Portal - Campaigns"
                ],
                "summary": "Preview campaign creative",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_campaigns_interfaces_http.CampaignPreviewResponse"
                        }
                    }
                }
            }
        },
...
        "/intents/predict": {
            "post": {
                "security": [
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Loads stored/cached events for a user, calls ML, persists intent and reward when threshold is met.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SDK - Intents"
                ],
                "summary": "Predict user intent (sync)",
                "parameters": [
                    {
                        "description": "User id to evaluate",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_intents_interfaces_http.PredictIntentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_intents_application.PredictIntentResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_platform_http.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/skykin-platform_internal_platform_http.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_campaigns_interfaces_http.CampaignListResponse": {
            "type": "object",
            "properties": {
                "campaigns": {
                    "type": "array",
                    "items": {}
                }
            }
        },
        "internal_campaigns_interfaces_http.CampaignPreviewResponse": {
            "type": "object",
            "properties": {
                "campaign_name": {
                    "type": "string"
                },
                "channel_label": {
                    "type": "string"
                },
                "format": {
                    "type": "string"
                },
                "preview": {
                    "type": "object",
                    "additionalProperties": true
                },
                "simulator": {
                    "type": "boolean"
                }
            }
        },
        "internal_campaigns_interfaces_http.CreateCampaignRequest": {
            "type": "object",
            "required": [
                "application_id",
                "creative_format",
                "destination_url",
                "name",
                "target_intent"
            ],
            "properties": {
                "application_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "body_text": {
                    "type": "string",
                    "example": "Trade crypto with zero fees today"
                },
                "canvas_json": {
                    "type": "object",
                    "additionalProperties": true
                },
                "creative_format": {
                    "type": "string",
                    "example": "BANNER"
                },
                "daily_budget_cap": {
                    "type": "number",
                    "example": 100
                },
                "destination_url": {
                    "type": "string",
                    "example": "https://example.com/offer"
                },
                "image_url": {
                    "type": "string",
                    "example": "https://cdn.example.com/banner.png"
                },
                "name": {
                    "type": "string",
                    "example": "Crypto Promo"
                },
                "target_intent": {
                    "type": "string",
                    "example": "crypto_interest"
                },
                "title": {
                    "type": "string",
                    "example": "Save on fees"
                },
                "total_budget_cap": {
                    "type": "number",
                    "example": 1000
                }
            }
        },
...
        "internal_intents_interfaces_http.PredictIntentRequest": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "string",
                    "example": "user_test_batch_001"
                }
            }
        },
...
        "skykin-platform_internal_campaigns_model.Campaign": {
            "type": "object",
            "properties": {
                "advertiserID": {
                    "type": "string"
                },
                "applicationID": {
                    "type": "string"
                },
                "bodyText": {
                    "type": "string"
                },
                "canvasJSON": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "creativeFormat": {
                    "type": "string"
                },
                "dailyBudgetCap": {
                    "type": "number"
                },
                "destinationURL": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageURL": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "targetIntent": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "totalBudgetCap": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                },
                "validationNotes": {
                    "type": "string"
                },
                "validationStatus": {
                    "type": "string"
                }
            }
        },
        "skykin-platform_internal_intents_application.PredictIntentResult": {
            "type": "object",
            "properties": {
                "confidence": {
                    "type": "number"
                },
                "intent": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "reward": {
                    "$ref": "#/definitions/skykin-platform_internal_intents_application.Reward"
                },
                "reward_triggered": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                },
                "top_signals": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "skykin-platform_internal_intents_application.Reward": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "reward_id": {
                    "type": "string"
                },
                "reward_type": {
                    "type": "string"
                }
            }
        },
        "skykin-platform_internal_platform_http.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "description": "SDK publishable key (pk_live_...)",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Enter your JWT token as: Bearer \u003ctoken\u003e (developer portal or ad portal)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}